refactored Metadata.java and corresponding tests to snippet format
